refactor(board): :recycle: Rename page board (boards)
